stream.py: include logo in video
Introduces new flags to include a logo by an URL into the video. The
logo will be placed in a corner, by default the top right one, as this
does not collide with the list of current speakers. However, one can
override this, e.g., when the bar of current speakers was disabled.

The logo's style attribute places it in the background and limits its
height to a tenth of the visible screen size. This prevents distorting
other elements, e.g., when lots of webcams are in use.