stream.py: shrink navbar when empty
Followup to #153. When both the meeting tile as well as the list of
current speakers are hidden, the top navbar becomes empty. However, it
still blocks another 112px for an empty box. As all elements still have
a margin, removing this forced height will not place videos directly
under the top.